feat(stm32c5): shared CAN ownership model and restructured test suite
modcan.c:
- Replace we_started flag with shared device-state + refcount: a CAN
  controller is owned by a shared slot; multiple can_obj_t may bind to it
  and can_stop runs only when refcount reaches 0 (fixes deinit/GC of one
  object taking down a controller still used by another).
- First owner applies config + starts; later owners validate fd / bitrate /
  loopback / data-bitrate match and reject on mismatch without touching the
  running controller or refcount.
- RX queue 128 -> 32 frames (9.3KB -> 2.4KB per object) to fit the
  fragmented 105KB heap; CAN_MAX_DEVICES=4 defensive cap.
- Per-object dual RX filters (std + ext) so extended frames reach the msgq.
- send(): validate id range and set CAN_FRAME_IDE for 29-bit ids.
- Idempotent deinit (safe under double-deinit / GC after explicit deinit).

Tests:
- xiao_stm32c5_full_test.py: refactor to single combined entry point
  (base peripherals + 9 FDCAN tests); add owner / mismatch / id CAN tests.
- xiao_stm32c5_base_test.py (new): granular base-peripheral suite.
- xiao_stm32c5_can_test.py (new): 9 FDCAN loopback tests (PR#22 fixes).
- can_interconnect.py: two-board interconnect refinements.

Docs:
- README + release notes: full_test is the single entry point; manifest
  comment notes base/can/interconnect scripts are not frozen.
